Refactor chat normalizes to their own directories
