+++ mlflow/server/js/src/shared/web-shared/model-trace-explorer/ModelTrace.types.ts
@@ -365,37 +365,8 @@ export type ModelTraceChatResponse = {
   }[];
 };
 
-export type LlamaIndexChatResponse = {
-  message: ModelTraceChatMessage;
-};
-
 export type ModelTraceChatInput = {
   messages: RawModelTraceChatMessage[];
-};
-
-// it has other fields, but we only care about these for now
-export type LangchainBaseMessage = {
-  content?: string;
-  type: 'human' | 'user' | 'assistant' | 'ai' | 'system' | 'tool' | 'function';
-  tool_calls?: LangchainToolCallMessage[];
-  tool_call_id?: string;
-  additional_kwargs?: {
-    // some langchain models have tool_calls specified in additional_kwargs in
-    // OpenAI format. this appears to be a bug, but we should still try to handle it
-    tool_calls?: ModelTraceToolCall[];
-  };
-};
-
-export type LangchainToolCallMessage = {
-  name: string;
-  // an object with the arguments to the tool call.
-  // should be stringified before display.
-  args: any;
-  id: string;
-};
-
-export type LangchainChatGeneration = {
-  message: LangchainBaseMessage;
 };
 
 export type AssessmentSourceType = 'SOURCE_TYPE_UNSPECIFIED' | 'HUMAN' | 'LLM_JUDGE' | 'CODE';
@@ -472,151 +443,3 @@ export interface ExpectationAssessment extends AssessmentBase {
 }
 
 export type Assessment = FeedbackAssessment | ExpectationAssessment;
-
-export type GeminiChatInput = {
-  contents: string | GeminiContent[];
-};
-
-export type GeminiChatOutput = {
-  candidates: GeminiCandidate[];
-  // propmtFeedback: GeminiPropmptFeedback;
-  // usageMetadata: GeminiUsageMetadata;
-  modelVersion: string;
-  responseId: string;
-};
-
-export type GeminiCandidate = {
-  content: GeminiContent;
-  finishReason:
-    | 'FINISH_REASON_UNSPECIFIED'
-    | 'STOP'
-    | 'MAX_TOKENS'
-    | 'SAFETY'
-    | 'RECITATION'
-    | 'LANGUAGE'
-    | 'OTHER'
-    | 'BLOCKLIST'
-    | 'PROHIBITED_CONTENT'
-    | 'SPII'
-    | 'MALFORMED_FUNCTION_CAL'
-    | 'IMAGE_SAFETY'
-    | 'UNEXPECTED_TOOL_CAL';
-  // safetyRatings: GeminiSafetyRating[]
-  // citationMetadata: GeminiCitationMetadata
-  // tokenCount: number
-  // groundingAttributions: GeminiGroundingAttribution[]
-  // groundingMetadata: GeminiGroundingMetadata
-  // avgLogprobs: number
-  // logprobsResult: GeminiLogprobsResult
-  // urlContextMetadata: GeminiUrlContextMetadata
-  // index: number
-};
-
-export type GeminiContent = {
-  role: 'user' | 'model';
-  parts: GeminiContentPart[];
-};
-
-export type GeminiContentPart = { text: string };
-// | { inlineData: GeminiBlob }
-// | { functionCall: GeminiFunctionCall }
-// | { functionResponse: GeminiFunctionResponse }
-// | { fileData: GeminiFileData }
-// | { executableCode: GeminiExecutableCode }
-// | { codeExecutionResult: GeminiCodeExecutionResult };
-
-// type GeminiBlob = {
-//   mimeType: string;
-//   data: string;
-// };
-
-// type GeminiFunctionCall = {
-//   id: string;
-//   name: string;
-//   args: Record<string, string>;
-// };
-
-// type GeminiFunctionResponse = {
-//   id: string;
-//   name: string;
-//   response: Record<string, string>;
-//   willContinue: boolean;
-//   scheduling: 'SCHEDULING_UNSPECIFIED' | 'SILENT' | 'WHEN_IDLE' | 'INTERRUPT';
-// };
-
-// type GeminiFileData = {
-//   mimeType: string;
-//   fileUri: string;
-// };
-
-// type GeminiExecutableCode = {
-//   language: 'LANGUAGE_UNSPECIFIED' | 'PYTHON';
-//   code: string;
-// };
-
-// type GeminiCodeExecutionResult = {
-//   outcome: 'OUTCOME_UNSPECIFIED' | 'OUTCOME_OK' | 'OUTCOME_FAILED' | 'OUTCOME_DEADLINE_EXCEEDED';
-//   output: string;
-// };
-
-export type AnthropicMessagesInput = {
-  messages: AnthropicMessageParam[];
-};
-
-export type AnthropicMessagesOutput = {
-  id: string;
-  content: AnthropicContentBlock[];
-  role: 'assistant';
-  type: 'message';
-  // model: Model;
-  // stop_reason: StopReason | null;
-  // stop_sequence: string | null;
-  // usage: Usage;
-};
-
-type AnthropicContentBlock = AnthropicTextBlock;
-// | ThinkingBlock
-// | RedactedThinkingBlock
-// | ToolUseBlock
-// | ServerToolUseBlock
-// | WebSearchToolResultBlock;
-
-export type AnthropicMessageParam = {
-  content: string | AnthropicContentBlockParam[];
-  role: 'user' | 'assistant';
-};
-
-export type AnthropicContentBlockParam = AnthropicTextBlockParam | AnthropicImageBlockParam;
-// | DocumentBlockParam
-// | ThinkingBlockParam
-// | RedactedThinkingBlockParam
-// | ToolUseBlockParam
-// | ToolResultBlockParam
-// | ServerToolUseBlockParam
-// | WebSearchToolResultBlockParam;
-
-type AnthropicTextBlockParam = {
-  text: string;
-  type: 'text';
-};
-
-type AnthropicTextBlock = {
-  text: string;
-  type: 'text';
-};
-
-type AnthropicImageBlockParam = {
-  source: AnthropicBase64ImageSource | AnthropicURLImageSource;
-  type: 'image';
-};
-
-type AnthropicBase64ImageSource = {
-  type: 'base64';
-  data: string;
-  media_type: 'image/jpeg' | 'image/png' | 'image/gif' | 'image/webp';
-};
-
-type AnthropicURLImageSource = {
-  type: 'url';
-  url: string;
-};
